XML value configuration and disable date which not include in selecte…

+++ CalendarView/src/main/java/com/shahzadafridi/calendarview/MonthConfiguration.kt
@@ -5,5 +5,5 @@ data class MonthConfiguration(
     var mSelectedClr: Int? = null,
     var mUnSelectedClr: Int? = null,
     var mFont: Int? = null,
-    var mTxtSize: Int? = null,
+    var mTxtSize: Float? = null,
 )
